oop and abs all the way down
