+++ tests/test_shapes.py
@@ -3,7 +3,7 @@ from contextlib import redirect_stdout
 
 import pytest
 
-from word_search_generator import PuzzleNotGeneratedError, WordSearch
+from word_search_generator.game.word_search import PuzzleNotGeneratedError, WordSearch
 from word_search_generator.mask import shapes
